Update wampf_naive_cbs.h
There wasn't any check on local_path.actions.second, so, in most cases, the goal_idx was out of the index range for that specific path, causing the program to terminate.

I added a simple 'if' statement to check whether this condition is true or not. In the event that goal_idx exceeds the bounds, I use the last timestamp of that local path instead.